requirements.txt
# requirements.txt
pytest==8.3.0
playwright==1.45.0
requests==2.32.0
allure-pytest==2.13.5